requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
mistralai==0.0.12
openai==1.12.0
python-dotenv==1.0.1
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0 